+++ spec/rails/assets/manifest/manifest_spec.rb
@@ -1,22 +1,59 @@
 # frozen_string_literal: true
 
 require 'tempfile'
+require 'pathname'
 
 RSpec.describe Rails::Assets::Manifest::Manifest do
-  subject(:manifest) { described_class.new(path: path) }
+  subject(:manifest) { described_class.new(files: tmp.join('*.json')) }
 
-  let(:tmpfile) { Tempfile.new 'manifest' }
-  let(:path) { tmpfile.path }
+  let(:tmp) { Pathname.new(Dir.mktmpdir('manifest')) }
+  let(:tmpfile) { tmp.join('manifest.json').open('w') }
 
-  after { tmpfile.unlink }
+  after { tmp.rmtree }
+
+  let(:payload) do
+    {'app.js': 'app-digest.js'}
+  end
+
+  before do
+    tmp.join('manifest.json').open('w') {|f| f.write JSON.generate(payload) }
+  end
+
+  context 'with multiple files in glob pattern' do
+    before do
+      tmp.join('second-manifest.json').open('w') do |file|
+        file.write JSON.generate('app2.js': 'app2-digest.js')
+      end
+    end
+
+    it 'includes entries from both files' do
+      expect(manifest).to be_key 'app.js'
+      expect(manifest).to be_key 'app2.js'
+    end
+
+    context 'with multiple patterns/files' do
+      subject(:manifest) do
+        described_class.new(files: [tmp.join('manifest.json'), tmp.join('second*.json')])
+      end
+
+      it 'includes entries from both files' do
+        expect(manifest).to be_key 'app.js'
+        expect(manifest).to be_key 'app2.js'
+      end
+    end
+  end
+
+  context 'with invalid entry' do
+    let(:payload) { {'app.js': { nosrc: true }} }
+    subject(:lookup) { manifest.lookup('app.js') }
+
+    it 'raises an error' do
+      expect { lookup }.to raise_error(::Rails::Assets::Manifest::ManifestInvalid)
+    end
+  end
 
   describe '#lookup' do
     subject(:entry) { manifest.lookup('app.js') }
-
-    before do
-      tmpfile.write JSON.dump('app.js': {src: 'app-digest.js'})
-      tmpfile.flush
-    end
 
     it 'returns manifest entry' do
       expect(entry).to be_a Rails::Assets::Manifest::Manifest::Entry