+++ spec/rails/assets/manifest_spec.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-
-require 'tempfile'
-
-RSpec.describe Rails::Assets::Manifest do
-end